fix(state): checked balance arithmetic and cap referrer fees to fare
Two money-path defects found in the clutch-node audit:

1. Balance math used `(balance as i64 + change) as u64` with no overflow
   check. Any balance above i64::MAX corrupted on credit — the genesis
   account (u64::MAX) collapsed to a tiny value on its first incoming
   transfer. Replaced with sign-aware checked_add/checked_sub (apply_delta);
   over/underflow is logged and clamped instead of wrapping.

2. RidePay computed `driver_amount = fare - total_deducted` as unchecked u64
   subtraction. referrer_fee_ceiling rounds up, so two referrers on a tiny
   fare (2% of 1 = 1 each) made the fees exceed the fare and underflow the
   driver amount — wrapping to ~u64::MAX in release (a money mint) and
   panicking in debug. Fees are now capped at the fare via split_fare(), and
   the ceiling is hardened against multiply overflow.

Genesis funding changes from u64::MAX to i64::MAX. Balance deltas travel as
i64, so u64::MAX only funded correctly by two's-complement wrap; checked math
requires the genesis balance to fit i64. i64::MAX (~9.2e18) is still
effectively unlimited for a testnet faucet.

BREAKING: the genesis transaction value changed, so the genesis block hash
changes. Nodes on this build are consensus-incompatible with chains created
before it — deploy against a fresh chain (reset the node DBs).

Unit tests cover apply_delta and split_fare; existing money-path integration
tests (transfer, block_reward, referrer_account, ride_sharing,
balance_effects) pass.

+++ src/node/transactions/transaction.rs
@@ -47,7 +47,11 @@ impl Transaction {
             0,
             FunctionCall::Transfer(Transfer {
                 to: "0xdeb4cfb63db134698e1879ea24904df074726cc0".to_string(),
-                value: u64::MAX,
+                // ponytail: i64::MAX, not u64::MAX. Balance deltas travel as i64
+                // (transfer.rs `value as i64`), so funding u64::MAX only ever worked by
+                // two's-complement wrap. i64::MAX (~9.2e18) is still effectively infinite
+                // for a testnet faucet and keeps every balance representable in i64.
+                value: i64::MAX as u64,
             }),
         );
